listed to websocket events from the SysAp
This improves the overall loading time of the UI and makes devices
react instantly. There is no delay anymore and state changes outside
HomeKit are pushed immediately to HomeKit to be processed.
Additionally this reduces load on the Busch-Jaeger API as by default
we only pull all data once every 60 seconds to make sure that it doesn't
drift off if we have missed some data (just in case)